backend/requirements.txt
fastapi
uvicorn
python-keycloak
python-jose[cryptography]
debugpy
PyJWT
peewee